refactor demos into modular pages
